CrowdFunding/requirements.txt
amqp==1.4.9
anyjson==0.3.3
autopep8==1.3.5
billiard==3.3.0.23
celery==3.1.26.post2
crispy==0.5.0
diff-match-patch==20121119
Django==2.0.5
django-celery==3.2.2
django-crispy-forms==1.7.2
django-formtools==2.1
django-pure-pagination==0.3.0
django-redis==4.9.0
django-reversion==2.0.13
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
itsdangerous==0.24
jdcal==1.4
kombu==3.0.37
openpyxl==2.5.3
Pillow==5.1.0
pycodestyle==2.4.0
pycryptodomex==3.6.1
PyMySQL==0.8.1
python-alipay-sdk==1.7.1
pytz==2018.4
PyYAML==3.12
redis==2.10.6
six==1.11.0
unicodecsv==0.14.1
vine==1.1.4
xadmin==2.0.1
xlrd==1.1.0
xlwt==1.3.0
